discord-bot/requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
py-cord
requests
python-escpos[usb]
python-dotenv
Pillow
fastapi
uvicorn[standard]
python-multipart
qrcode[pil]
markdown
